Reworked telegram service
