+++ service/src/main/java/greencity/config/TelegramBotConfig.java
@@ -2,7 +2,9 @@ package greencity.config;
 
 import greencity.service.ubs.TelegramService;
 import greencity.ubstelegrambot.UBSTelegramBot;
+import greencity.ubstelegrambot.service.TelegramExecutor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
@@ -20,8 +22,8 @@ public class TelegramBotConfig {
     private String botToken;
 
     @Bean
-    public UBSTelegramBot ubsTelegramBot(TelegramService telegramService) {
-        return new UBSTelegramBot(botToken, botName, telegramService);
+    public UBSTelegramBot ubsTelegramBot(TelegramService telegramService, ApplicationContext applicationContext, TelegramExecutor executor) {
+        return new UBSTelegramBot(botToken, botName, telegramService, applicationContext, executor);
     }
 
     @Bean
